==> 00-json-to-dataframe.ipynb <==
# Data Transformation

Neste primeiro notebook, vamos inspecionar os dados disponibilizados pelo Spotify por meio do [Recsys Challenge 2018](https://recsys-challenge.spotify.com/) e transformá-los de sua estrutura atual em arquivos json para uma estrutura tabular pronta para processamentos seguintes.

@author: [Pedro Correia](https://github.com/pfcor)


import json
import os
import codecs
import datetime as dt
import pandas as pd
import tqdm
import mpd

Para iniciar as investigações, vamos observar o conteúdo (limpo e organizado) de uma playlist aleatória:

mpd.show_playlist(433)

Está presente na playlist acima todas as informações de fato disponíveis no dataset. Associada a cada uma das 1000000 de playlists temos o seu título, a quantidade de seguidores, a data de última edição e a quantidade de edições, se a playlist é colaborativa ou não, além das quantidades de artistas, faixas e de minutos de duração.

<hr>

O próximo passo é organizar essas informações em formato tabular, para utilização na sequência do trabalho.

def tabulate_playlist_data(input_path, output_path):
    
    playlist_features = [
        'pid',
        'name',
        'collaborative',  
        'modified_at', 
        'duration_ms',
        'num_tracks',
        'num_albums', 
        'num_artists',
        'num_followers',
        'num_edits'
    ]
    
    track_features = [
        'track_uri',
        'track_name',
        'album_uri',
        'album_name',
        'artist_uri',
        'artist_name',
        'duration_ms'
    ]

    slice_files = [
        os.sep.join((input_path, slice_file)) 
        for slice_file in 
        sorted(os.listdir(f'{input_path}'))
        if slice_file.startswith('mpd.slice')
    ]
    
    tracks = set() # accumulator for all tracks that appear
    
    for i, slice_file in tqdm.tqdm_notebook(enumerate(slice_files), total=len(slice_files), desc="Data Files"):
        
        # setting up accumulators
        playlist_metadata = [] # aggregating info about the playlist (name, num edits, num artists, duration...)
        track_metadata    = [] # aggregating info about the track (title, album, artist...)
        playlist_track    = [] # to-from playlist/track. also registering position of the track on the playlist here
        
        # reading file with playlists dataii
        with codecs.open(slice_file, 'r', 'utf-8') as f:
            playlists = json.loads(f.read())['playlists']
        
        # processing each playlist into tabular data
        for playlist in playlists:
            
            # appending lists with each of the relevant fields
            playlist_metadata.append([playlist[feat] for feat in playlist_features])
            
            # processing the tracks array of the current playlist
            for track in playlist['tracks']:
                
                # adding the connection of each playlist with each track
                playlist_track.append([playlist['pid'], track['track_uri'], track['pos']])
                
                # building a database of all tracks we see
                if track['track_uri'] not in tracks:
                    
                    # appending the info listed above
                    track_metadata.append([track[feat] for feat in track_features])
                    
                    # keeping record of all tracks already seen
                    tracks.add(track['track_uri'])
        
        # playlist metadata
        playlists_df = pd.DataFrame(playlist_metadata, columns=playlist_features)
        playlists_df['collaborative'] = playlists_df['collaborative'].map({'false': 0, 'true': 1})
        playlists_df.to_csv(os.sep.join((output_path, "playlists.csv")), index=False, header=(i==0), mode="w" if (i==0) else "a")
        
        # track data
        tracks_df = pd.DataFrame(track_metadata, columns=track_features)
        tracks_df.to_csv(os.sep.join((output_path, "tracks.csv")), index=False, header=(i==0), mode="w" if (i==0) else "a")
        
        # playlist-track
        playlist_track_df = pd.DataFrame(playlist_track, columns=['pid', 'track_uri', 'pos'])
        playlist_track_df.to_csv(os.sep.join((output_path, "playlist_track.csv")), index=False, header=(i==0), mode="w" if (i==0) else "a")


# calling the function to create the tabular files
input_path  = "data/original"
output_path = "data/formatted"

tabulate_playlist_data(input_path, output_path)

Vamos agora brevemente inspecionar os arquivos que acabamos de criar:

> playlists

df_playlists = pd.read_csv("data/formatted/playlists.csv")

df_playlists.shape

Como esperado, são um milhão de playlists

df_playlists.head()

As informações estão registradas conforme esperado. Talvez caiba na sequência transformar os dados de `modified_at`e `duration_ms` em formatos de mais fácil entendimento, mas para o momento nossos objetivos foram alcançados.

# limpando namespace
del df_playlists

> tracks

aqui, vamos fazer algumas passos adicionais. vamos criar um id inteiro para as faixas e criar um dataset específico para os artistas:

df_tracks = pd.read_csv("data/formatted/tracks.csv")

df_tracks.shape

df_tracks.head(3).T

# criando dataset de artistas
df_artists = df_tracks.drop_duplicates("artist_uri", keep="first")[["artist_uri", "artist_name"]]
df_artists = df_artists.reset_index().rename(columns={"index": "aid"})
df_artists.head()

df_artists.to_csv("data/formatted/artists.csv", index=False)

df_tracks = df_tracks.reset_index().rename(columns={"index": "tid"})
df_tracks = df_tracks.merge(df_artists[["aid", "artist_uri"]], on="artist_uri", how="left")

df_tracks.sample(3).T

df_artists[df_artists["aid"]==108788]

df_tracks.to_csv("data/formatted/tracks.csv", index=False)

# limpando namespace
del df_tracks

> playlist-track

df_pt = pd.read_csv("data/formatted/playlist_track.csv")

df_pt.shape

df_tracks = df_tracks[["track_uri", "tid", "aid"]]

df_pt = df_pt.merge(df_tracks[["track_uri", "tid", "aid"]], on="track_uri", how="left").drop("track_uri", axis=1)[["pid", "tid", "aid", "pos"]]

df_pt.sample(5)

df_pt.to_csv("data/formatted/playlist_track.csv", index=False)

top_ten_tracks = df_pt["tid"].value_counts().head(20).reset_index().rename(columns={"index": "tid", "tid": "total"})
df_tracks[["tid", "track_name", "artist_name"]].merge(top_ten_tracks, on="tid").sort_values("total", ascending=False)

<hr>
