==> 01a-metadata-scraping-tracks.ipynb <==
# Metadata Scraping - Tracks

Os dados fornecidos pelo Spotify são bastante secos em termos de catergorização das músicas presentes em cada uma das playlists. Em uma estratégia utilizando puramente técnicas de Filtro Colaborativo, que é nosso foco neste projeto, isso seria suficiente. Mas buscaremos observar também técnicas relacionadas sistemas Baseados em Conteúdo, bem como estratégias híbridas, bem como será interessante analisar alguns padrões nas diversas listagens disponíveis.

Assim, no presente notebook, nos utilizamos do módulo [Spotipy](https://spotipy.readthedocs.io/en/latest/) para nos utilizar de dados do Spotify para enriquecer nossa base de dados.

Dados de <b>FAIXAS</b>

<br><br>
<hr>
@author: [Pedro Correia](https://github.com/pfcor)

import math
import pprint
from collections import defaultdict
import pandas as pd
import numpy as np
import tqdm
import time
import datetime as dt
import os

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

Vamos iterar faixa a faixa e obter informações relacionadas ao artista (estilos, popularidade, número de seguidores e artistas similares)e à música (níveis de ritmos, instrumentação, etc)

def spotipy_auth():
    
    # initializing and authenticating
    from spotipy_credencials import CLIENT_ID, CLIENT_SECRET
    ccm = SpotifyClientCredentials( client_id=CLIENT_ID, client_secret=CLIENT_SECRET )
    sp = spotipy.Spotify(client_credentials_manager=ccm)
    return sp


def get_tracks_uri(ntracks=None):
    
    return pd.read_csv("data/formatted/tracks.csv", nrows=ntracks, usecols=["track_uri"])["track_uri"].values


def get_top_n_related_artists(artist_uri, n=3, sort_by="popularity", include_name=False):
    
    assert sort_by in ["popularity", "similarity"]
    
    related_artists = [
        (rartist["uri"], rartist["name"], rartist["popularity"]) 
        for rartist 
        in sp.artist_related_artists(artist_uri)["artists"]
    ]
    
    if sort_by=="popularity":
        related_artists = sorted(related_artists, key=lambda x: -x[-1])[:n]
    
    return [rartist[:2] if include_name else rartist[0] for rartist in related_artists[:n]]

def scrape_track_metadata(batch_size=50, output_folder="data/metadata", start_batch=0):
    
    if not os.path.exists(output_folder):
        os.makedirs(output_folder)
    filename = output_folder + '/tracks_metadata.csv'
    failed_index = []
    
    
    print("|Initializing Spotify API...", end=" ", flush=True)
    sp = spotipy_auth()
    print("ok\n")
    
    print("|Loading tracks...", end=" ", flush=True)
    tracks = get_tracks_uri()
    n_tracks = len(tracks)
    print("ok")
    
    print(f"    > {n_tracks} tracks to be processed\n")
    
    print("|Gathering Spotify data")
    t0 = time.time()
    

    batches_count = math.ceil(n_tracks / batch_size)
    print(f"    > {batches_count} batches to be processed")
    
    for i in range(start_batch, batches_count):
        spotify_data = defaultdict(list)
        
#         print(f'        - batch # {i+1:0>2}', end=" ", flush=True)
        try:
            i_0 = i * batch_size
            i_f = i_0 + batch_size           
            batch = tracks[i_0:i_f]

            # general data
            tracks_data = sp.tracks(batch)["tracks"]
            for track in tracks_data:
                if track != None:
                    spotify_data['track_uri'].append(track['uri'])
                    spotify_data['popularity'].append(track['popularity'])
                    spotify_data['realease_year'].append(int(track['album']['release_date'].split("-")[0]))
                else:
                    spotify_data['track_uri'].append(np.nan)
                    spotify_data['popularity'].append(np.nan)
                    spotify_data['realease_year'].append(np.nan)

            # audio data
            track_audio_features = sp.audio_features(batch)
            for track in track_audio_features:
                if track != None:
                    spotify_data['acousticness'].append(track['acousticness'])
                    spotify_data['danceability'].append(track['danceability'])
                    spotify_data['energy'].append(track['energy'])
                    spotify_data['instrumentalness'].append(track['instrumentalness'])
                    spotify_data['mode'].append(track['mode'])
                    spotify_data['loudness'].append(track['loudness'])
                    spotify_data['speechiness'].append(track['speechiness'])
                    spotify_data['tempo'].append(track['tempo'])
                    spotify_data['time_signature'].append(track['time_signature'])
                    spotify_data['valence'].append(track['valence'])

                else:
                    spotify_data['acousticness'].append(np.nan)
                    spotify_data['danceability'].append(np.nan)
                    spotify_data['energy'].append(np.nan)
                    spotify_data['instrumentalness'].append(np.nan)
                    spotify_data['mode'].append(np.nan)
                    spotify_data['loudness'].append(np.nan)
                    spotify_data['speechiness'].append(np.nan)
                    spotify_data['tempo'].append(np.nan)
                    spotify_data['time_signature'].append(np.nan)
                    spotify_data['valence'].append(np.nan)

            try:
                spotify_data = pd.DataFrame(spotify_data)
                spotify_data.to_csv(filename, sep=";", index=False, header=(i==0), mode="w" if (i==0) else "a")
            except:
                for feat, vals in spotify_data.items():
                    print(feat, len(vals))
                
            if ((i+1)%100 == 0) or ((i+1) == batches_count):
                print(f'        - {i+1:>8} batches completed | {batches_count-(i+1):>8} to go | elapsed time: {str(dt.timedelta(seconds=time.time()-t0))}')
        except TimeoutError:
            failed_index.append((i_0, i_f))
            sp = spotipy_auth()
        


scrape_track_metadata(start_batch=45200)

tracks_metadata = pd.read_csv("data/metadata/tracks_metadata.csv", sep=";")

tracks_metadata.shape

tracks_metadata.head()

tracks = pd.read_csv("data/formatted/tracks.csv", usecols=["tid", "track_uri", "track_name", "artist_name"])

tracks_metadata.head().merge(tracks, on="track_uri", how="left").T

tracks_metadata.loc[tracks_metadata["energy"]==1, ["track_uri", "energy"]].merge(tracks, on="track_uri", how="left")

<hr>
