==> 01b-metadata-scraping-artists.ipynb <==
# Metadata Scraping - Artists

Os dados fornecidos pelo Spotify são bastante secos em termos de catergorização das músicas presentes em cada uma das playlists. Em uma estratégia utilizando puramente técnicas de Filtro Colaborativo, que é nosso foco neste projeto, isso seria suficiente. Mas buscaremos observar também técnicas relacionadas sistemas Baseados em Conteúdo, bem como estratégias híbridas, bem como será interessante analisar alguns padrões nas diversas listagens disponíveis.

Assim, no presente notebook, nos utilizamos do módulo [Spotipy](https://spotipy.readthedocs.io/en/latest/) para nos utilizar de dados do Spotify para enriquecer nossa base de dados.

Dados de <b>ARTISTAS</b>
<br><br>
<hr>
@author: [Pedro Correia](https://github.com/pfcor)

import math
import pprint
from collections import defaultdict
import pandas as pd
import numpy as np
import tqdm
import time
import datetime as dt
import os

import spotipy
from spotipy.oauth2 import SpotifyClientCredentials

Vamos iterar faixa a faixa e obter informações relacionadas ao artista (estilos, popularidade, número de seguidores e artistas similares)e à música (níveis de ritmos, instrumentação, etc)

def spotipy_auth():
    
    # initializing and authenticating
    from spotipy_credencials import CLIENT_ID, CLIENT_SECRET
    ccm = SpotifyClientCredentials( client_id=CLIENT_ID, client_secret=CLIENT_SECRET )
    sp = spotipy.Spotify(client_credentials_manager=ccm)
    return sp


def get_artists_uri():
    
    return pd.read_csv("tab_data/tracks_metadata.csv", usecols=["artist_uri"])["artist_uri"].unique()


def get_top_n_related_artists(artist_uri, n=3, sort_by="popularity", include_name=False):
    
    assert sort_by in ["popularity", "similarity"]
    
    related_artists = [
        (rartist["uri"], rartist["name"], rartist["popularity"]) 
        for rartist 
        in sp.artist_related_artists(artist_uri)["artists"]
    ]
    
    if sort_by=="popularity":
        related_artists = sorted(related_artists, key=lambda x: -x[-1])[:n]
    
    return [rartist[:2] if include_name else rartist[0] for rartist in related_artists[:n]]

def scrape_artist_metadata(batch_size=50, output_folder="data/metadata"):
    
    print("|Initializing Spotify API...", end=" ", flush=True)
    sp = spotipy_auth()
    print("ok\n")
    
    print("|Loading artists...", end=" ", flush=True)
    artists = get_artists_uri()
    n_artists = len(artists)
    print("ok")
    
    print(f"    > {n_artists} artists to be processed\n")
    
    print("|Gathering Spotify data")
    t0 = time.time()
    
    spotify_data = defaultdict(list)

    batches_count = math.ceil(n_artists / batch_size)
    print(f"    > {batches_count} batches to be processed")
    
    for i in range(batches_count):

        i_0 = i * batch_size
        i_f = i_0 + batch_size           
        batch = artists[i_0:i_f]
        
        # general data
        artists_data = sp.artists(batch)["artists"]
        
        for artist in artists_data:
        
            if artist != None:
                spotify_data['artist_uri'].append(artist['uri'])
                spotify_data['popularity'].append(artist['popularity'])
                spotify_data['genres'].append(artist["genres"])
#                 spotify_data["related_artists"].append(get_top_n_related_artists(artist["uri"]))
            else:
                spotify_data['artist_uri'].append(np.nan)
                spotify_data['popularity'].append(np.nan)
                spotify_data['genres'].append(np.nan)
#                 spotify_data["related_artists"].append(np.nan)
          
        if ((i+1)%100 == 0) or ((i+1) == batches_count):
            print(f'        - {i+1:>8} batches completed | {batches_count-(i+1):>8} to go | elapsed time: {str(dt.timedelta(seconds=time.time()-t0))}')
    
    if not os.path.exists(output_folder):
        os.makedirs(output_folder)
    filename = output_folder + '/artists_metadata.csv'
    print(f"    > Saving to file: {filename}")
    spotify_data = pd.DataFrame(spotify_data)
    spotify_data.to_csv(filename, index=False, sep=";")


scrape_artist_metadata()

<hr>

artists_metadata = pd.read_csv("data/metadata/artists_metadata.csv", sep=";")

artists_metadata["genres"][0]

import ast

ast.literal_eval(artists_metadata["genres"][0])

from collections import Counter

genders_counter = Counter()

genders_counter = Counter()
for genres in artists_metadata["genres"].values:
    genders_counter.update(ast.literal_eval(genres))

genders_counter

genders_counter.most_common(30)
