==> 02-convert-to-feather.ipynb <==
# Convert To Feather

Vamos lidar com arquivos relativamente grandes. De forma a agilizar o processo de leitura dos dados na diversas iterações do projeto, vamos converter os arquivos atualmente em csv para feather.

<br><br>
<hr>
@author: [Pedro Correia](https://github.com/pfcor)

import pandas as pd
import feather as ft
import os

Para comparação, vamos primeiramente, apenas ler os arquivos csv e medir quanto tempo leva:

%%time

for csv_file in os.listdir("data/formatted/"):
    if not csv_file.endswith(".csv"):
        continue
    print(f"{f'reading {csv_file}...':<35}", end=" ", flush=True)
    df = pd.read_csv(os.path.join("data/formatted/", csv_file))
    print("ok")

Vamos agora convertê-los para feather e salvar o novo arquivo na mesma pasta:

%%time

for csv_file in os.listdir("data/formatted/"):
    if not csv_file.endswith(".csv"):
        continue
    print(f"{f'converting {csv_file}...':<35}", end=" ", flush=True)
    df = pd.read_csv(os.path.join("data/formatted/", csv_file))
    df.to_feather(os.path.join("data/formatted/", csv_file.replace(".csv", ".fthr")))
    print("ok")

Agora, vamos apenas ler os aruivos feather para medir o ganho de tempo em relação ao csv puro.

%%time

for fthr_file in os.listdir("data/formatted/"):
    if not fthr_file.endswith(".fthr"):
        continue
    print(f"{f'reading {fthr_file}...':<35}", end=" ", flush=True)
    df = ft.read_dataframe(os.path.join("data/formatted/", fthr_file))
    print("ok")

Cerca de 10x mais rápida leitura.

<hr>
